+++ crates/ui/src/error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum Error {
+    NoUpdate
+}
