+++ crates/ui/src/lib.rs
@@ -1,4 +1,5 @@
 mod debugger;
 mod style;
+mod error;
 
 pub use debugger::Debugger;
